event-detail改为复用
