+++ src/app/shared/components/event-detail/event-detail.component.html
@@ -1,0 +1,118 @@
+<div fxLayout="row">
+  <div fxFlex.lg="1 1 calc(1em)"></div>
+  <div fxFlex="8 8 calc(8em)" class="card">
+    <div class="card-header card-header-primary">
+      <h4 class="card-title">校园培训活动详情</h4>
+    </div>
+    <div class="card-body" *ngIf="isLoading">
+      <div class="detail-section" fxLayout="column" fxLayoutAlign="start stretch" fxLayoutGap.md="10px">
+        <div class="detail-item" fxLayout="row" fxLayout.xs="column" fxLayoutAlign="space-between start"
+          fxLayoutGap="15px">
+          <span class="item-title" fxFlex="1 0 20%">项目名称</span>
+          <div class="item-content" fxFlex="1 0 78%">
+            <p>{{  programName }}</p>
+          </div>
+          <div fxFlex="1 0 2%"></div>
+        </div>
+        <div class="detail-item" fxLayout="row" fxLayout.xs="column" fxLayoutAlign="space-between start"
+          fxLayoutGap="15px">
+          <span class="item-title" fxFlex="1 0 20%">项目单位</span>
+          <div class="item-content" fxFlex="1 0 78%">
+            <p>{{ department  }}</p>
+          </div>
+          <div fxFlex="1 0 2%"></div>
+        </div>
+        <div class="detail-item" fxLayout="row" fxLayout.xs="column" fxLayoutAlign="space-between start"
+          fxLayoutGap="15px">
+          <span class="item-title" fxFlex="1 0 20%">项目类别</span>
+          <div class="item-content" fxFlex="1 0 78%">
+            <p>{{ category }}</p>
+          </div>
+          <div fxFlex="1 0 2%"></div>
+        </div>
+        <div class="detail-item" fxLayout="row" fxLayout.xs="column" fxLayoutAlign="space-between start"
+          fxLayoutGap="15px">
+          <span class="item-title" fxFlex="1 0 20%">活动名称</span>
+          <div class="item-content" fxFlex="1 0 78%">
+            <p>{{ item.name }}</p>
+          </div>
+          <div fxFlex="1 0 2%"></div>
+        </div>
+        <div class="detail-item" fxLayout="row" fxLayout.xs="column" fxLayoutAlign="space-between start"
+          fxLayoutGap="15px">
+          <span class="item-title" fxFlex="1 0 20%">编号</span>
+          <div class="item-content" fxFlex="1 0 78%">
+            <p>{{ item.id }}</p>
+          </div>
+          <div fxFlex="1 0 2%"></div>
+        </div>
+        <div class="detail-item" fxLayout="row" fxLayout.xs="column" fxLayoutAlign="space-between start"
+          fxLayoutGap="15px">
+          <span class="item-title" fxFlex="1 0 20%">时间</span>
+          <div class="item-content" fxFlex="1 0 78%">
+            <p>{{ item.time | date:'y年M月d日 H:mm' }}</p>
+          </div>
+          <div fxFlex="1 0 2%"></div>
+        </div>
+        <div class="detail-item" fxLayout="row" fxLayout.xs="column" fxLayoutAlign="space-between start"
+          fxLayoutGap="15px">
+          <span class="item-title" fxFlex="1 0 20%">地点</span>
+          <div class="item-content" fxFlex="1 0 78%">
+            <p>{{ item.location }}</p>
+          </div>
+          <div fxFlex="1 0 2%"></div>
+        </div>
+        <div class="detail-item" fxLayout="row" fxLayout.xs="column" fxLayoutAlign="space-between start"
+          fxLayoutGap="15px">
+          <span class="item-title" fxFlex="1 0 20%">时长</span>
+          <div class="item-content" fxFlex="1 0 78%">
+            <p>{{ item.num_hours | number:"0.1-1"  }}</p>
+          </div>
+          <div fxFlex="1 0 2%"></div>
+        </div>
+        <div class="detail-item" fxLayout="row" fxLayout.xs="column" fxLayoutAlign="space-between start"
+          fxLayoutGap="15px">
+          <span class="item-title" fxFlex="1 0 20%">参与人数</span>
+          <div class="item-content" fxFlex="1 0 78%">
+            <p>{{ item.num_participants }}</p>
+          </div>
+          <div fxFlex="1 0 2%"></div>
+        </div>
+        <div class="detail-item" fxLayout="row" fxLayout.xs="column" fxLayoutAlign="space-between start"
+          fxLayoutGap="15px">
+          <span class="item-title" fxFlex="1 0 20%">截止时间</span>
+          <div class="item-content" fxFlex="1 0 78%">
+            <p>{{ item.deadline | date:'y年M月d日 H:mm' }}</p>
+          </div>
+          <div fxFlex="1 0 2%"></div>
+        </div>
+        <div class="detail-item" fxLayout="row" fxLayout.xs="column" fxLayoutAlign="space-between start"
+          fxLayoutGap="15px">
+          <span class="item-title" fxFlex="1 0 20%">登记数</span>
+          <div class="item-content" fxFlex="1 0 78%">
+            <p>{{ item.num_enrolled }}</p>
+          </div>
+          <div fxFlex="1 0 2%"></div>
+        </div>
+        <div class="detail-item" fxLayout="row" fxLayout.xs="column" fxLayoutAlign="space-between start"
+           fxLayoutGap="15px">
+           <span class="item-title" fxFlex="1 0 20%">描述</span>
+           <div class="item-content" fxFlex="1 0 78%" [innerHTML]="description">
+           </div>
+           <div fxFlex="1 0 2%"></div>
+       </div>
+        <div fxLayout="row" fxLayoutAlign="end stretch" fxLayoutGap="5px">
+          <ng-container *ngIf="eventDetailType==1;else admin">
+              <button mat-raised-button color="primary">取消报名</button>
+          </ng-container>
+          <ng-template #admin>
+              <button mat-raised-button color="primary">导出签到表</button>
+              <button mat-raised-button color="primary" (click)="navigateToChangeEvent()">修改</button>
+          </ng-template>
+          <button mat-raised-button color="primary" (click)="location.back()">返回</button>
+        </div>
+      </div>
+    </div>
+  </div>
+  <div fxFlex.lg="1 1 calc(1em)"></div>
+</div>
