+++ src/app/shared/enums/event-detaile-type.enum.ts
@@ -1,0 +1,5 @@
+/** Enums of event detail types. */
+export enum EventDetailType {
+    ADMIN = 0,
+    USER = 1,
+}
