+++ src/app/user/modules/events/components/campus-event-detail/campus-event-detail.component.ts
@@ -1,55 +1,15 @@
-import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
-import { Location } from '@angular/common';
-import { CampusEvent } from 'src/app/shared/interfaces/event';
-import { DomSanitizer } from '@angular/platform-browser';
-import { map, switchMap } from 'rxjs/operators';
-import { ProgramService } from 'src/app/shared/services/programs/program.service';
-import { DepartmentService } from 'src/app/shared/services/department.service';
+import { Component } from '@angular/core';
 
+import { EventDetailType } from 'src/app/shared/enums/event-detaile-type.enum';
 
 @Component({
   selector: 'app-campus-event-detail',
   templateUrl: './campus-event-detail.component.html',
   styleUrls: ['./campus-event-detail.component.css']
 })
-export class CampusEventDetailComponent implements OnInit {
-
-  item: CampusEvent;
-  department: string;
-  category: string;
-  programName: string;
-  isLoading = false;
-
-  constructor(
-    private readonly route: ActivatedRoute,
-    private readonly sanitizer: DomSanitizer,
-    private readonly programService: ProgramService,
-    private readonly departmentService: DepartmentService,
-    readonly location: Location,
-  ) { }
-
-  ngOnInit() {
-    this.route.data.pipe(
-      switchMap(data => {
-        this.item = data.item;
-        return this.programService.getProgram(Number(this.item.program));
-      }),
-      switchMap(program => {
-        this.programName = program.name;
-        this.category = program.category_str;
-        return  this.departmentService.getDepartment(program.department);
-      }),
-      map(deparment => {
-        this.department = deparment.name;
-        this.isLoading = true;
-      })
-    ).subscribe();
-  }
-
-  get description() {
-    // Through this getter, we can bypass sanitizing and get raw HTML.
-    return this.sanitizer.bypassSecurityTrustHtml(this.item.description || /* istanbul ignore next */ '');
-  }
-
+export class CampusEventDetailComponent {
+  eventDetailType: EventDetailType = EventDetailType.USER;
+  constructor() { }
 }
+
+
